ex 4.15 to 4.22

+++ part4/blog-list/controllers/blogs.js
@@ -1,8 +1,11 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogRouter.get('/', async (req, res) => {
-  const blogs = await Blog.find({})
+  console.log(req.token)
+  const blogs = await Blog.find({}).populate('user', { name: 1, username: 1 })
   return res.json(blogs)
 })
 
@@ -17,14 +20,23 @@ blogRouter.get('/:id', async(req, res) => {
 
 blogRouter.post('/', async (req, res) => {
   const body = req.body
-  //const blog = new Blog(req.body)
+  const token = req.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if(!token || !decodedToken.id) {
+    return res.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
+  const user = await User.findById(decodedToken.id)
+  console.log(user)
   const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes || 0
+    ...req.body,
+    likes: body.likes || 0,
+    user: user._id
   })
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
   res.status(201).json(savedBlog)
 })
 
@@ -43,6 +55,14 @@ blogRouter.put('/:id', async(req, res) => {
 
 //delete req
 blogRouter.delete('/:id', async(req, res) => {
+  const token = req.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  const blog = await Blog.findById(req.params.id)
+  if (!token || !decodedToken.id || !blog || blog.user.toString() !== decodedToken.id) {
+    return res.status(401).json({
+      error: 'missing token or invalid'
+    })
+  }
   await Blog.findByIdAndRemove(req.params.id)
   return res.status(204).end()
 })
